[wip]migrate: to nx (#921)
* [wip]migrate: to nx

* fix: command

* chore: remove plugins

* fix: add handle error

* refactor: nx workspace

* fix: commands

+++ scripts/build/index.ts
@@ -1,30 +1,39 @@
 import { run } from '../shared/repository';
 import { getUseMockServer } from './env';
+import type { ChildProcess } from 'child_process';
 
 const build = async () => {
-  // schema
-  await run('@mya-ake-com/graphql-schema', ['build']);
+  let serverPs: ChildProcess | undefined = undefined;
+  try {
+    // schema
+    await run('@mya-ake-com/graphql-schema', ['build']);
 
-  // mock
-  await run('@mya-ake-com/mock', ['build']);
+    // mock
+    await run('@mya-ake-com/mock', ['build']);
 
-  // parser
-  await run('@mya-ake-com/parser', ['build']);
+    // parser
+    await run('@mya-ake-com/parser', ['build']);
 
-  // error
-  await run('@mya-ake-com/error', ['build']);
+    // error
+    await run('@mya-ake-com/error', ['build']);
 
-  // server
-  await run('@mya-ake-com/server', ['build']);
+    // server
+    await run('@mya-ake-com/server', ['build']);
 
-  const startCommand = getUseMockServer() ? 'start:mock' : 'start';
-  const serverPs = await run('@mya-ake-com/server', [startCommand], {
-    pararel: true,
-  });
+    const startTarget = getUseMockServer() ? 'start-mock' : 'start';
+    serverPs = await run('@mya-ake-com/server', [startTarget], {
+      pararel: true,
+    });
 
-  await run('@mya-ake-com/client', ['build']);
-
-  serverPs.kill();
+    await run('@mya-ake-com/client', ['build']);
+  } finally {
+    if (serverPs) {
+      serverPs.kill();
+    }
+  }
 };
 
-build();
+build().catch((err) => {
+  console.error(err);
+  process.exit(1);
+});
